requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
CRC-ITU==0.3.3
Flask==1.1.1
Flask-Session==0.3.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
libscrc==0.1.5
line-bot-sdk==1.0.2
MarkupSafe==1.1.1
Pillow==6.1.0
pypromptpay==0.4
qrcode==6.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
